Modified validators to prevent entering only white spaces in mandatory fields
